Lots of changes:
- Exception no longer produces an error message by default, only when it is left unhandled

- the DWI::Tractography::File & DWI::Tractography::ROI classes now build without errors

- added DataSet::Buffer & DataSet::Subset template classes

- small bug fix in Thread::Queue that causes double deletion of items

+++ src/dwi/tractography/common_parameters.h
@@ -45,9 +45,9 @@ namespace MR {
           void        reset ();
 
           Method      method;
-          std::string      cmd, source, mask;
+          std::string cmd, source, mask;
           float       step_size, max_dist, threshold, init_threshold, min_curv, mask_threshold;
-          uint32_t     max_num_tracks, max_num_tracks_generated, num_tracks_generated;
+          size_t      max_num_tracks, max_num_tracks_generated, num_tracks_generated;
 
           bool        unidirectional;
           Point       init_dir;
